+++ app/Http/Controllers/Badges/BadgeMakerController.php
@@ -13,6 +13,7 @@ use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
 use Illuminate\Validation\Rule;
+use Setting;
 use Validator;
 
 class BadgeMakerController extends Controller
@@ -172,6 +173,8 @@ class BadgeMakerController extends Controller
         $badgeCreator = new BadgeCreator($persons);
         if ($request->hasFile('alt_logo')) {
             $badgeCreator->logo = $request->file('alt_logo');
+        } else if (Setting::has('badges.logo_file')) {
+            $badgeCreator->logo = Storage::path(Setting::get('badges.logo_file'));
         }
         try {
             $badgeCreator->createPdf($title);