+++ app/Http/Controllers/Settings/SettingsController.php
@@ -19,6 +19,7 @@ class SettingsController extends Controller
             'bank' => __('bank.bank'),
             'shop' => __('shop.shop'),
             'library' => __('library.library'),
+            'badges' => __('badges.badges'),
         ];
     }
 
@@ -35,6 +36,7 @@ class SettingsController extends Controller
         'shop.help_article' => \App\Settings\Shop\HelpArticle::class,
         'library.default_lending_duration_days' => \App\Settings\Library\DefaultLendingDurationDays::class,
         'library.max_books_per_person' => \App\Settings\Library\MaxBooksPerPerson::class,
+        'badges.logo_file' => \App\Settings\Badges\LogoFile::class,
     ];
 
     private static function getSettings(): Collection